refactored position of merge-toolbar
