+++ src/components/input-panel/compiled-spec-header/index.tsx
@@ -1,14 +1,10 @@
 import stringify from 'json-stringify-pretty-compact';
 import * as React from 'react';
-import { ArrowDownCircle, ArrowUpCircle, ChevronDown, ChevronUp } from 'react-feather';
+import { ChevronDown, ChevronUp } from 'react-feather';
 import { connect } from 'react-redux';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
-import ReactTooltip from 'react-tooltip';
 import { bindActionCreators, Dispatch } from 'redux';
-import { mergeDeep } from 'vega-lite/build/src/util';
 import * as EditorActions from '../../../actions/editor';
-import { Mode, SIDEPANE } from '../../../constants';
-import ConfigEditorHeader from '../../config-editor/config-editor-header';
 
 const toggleStyle = {
   cursor: 'pointer',
@@ -20,58 +16,6 @@ class CompiledSpecDisplayHeader extends React.PureComponent<Props> {
   constructor(props) {
     super(props);
     this.editVegaSpec = this.editVegaSpec.bind(this);
-  }
-
-  public handleMergeConfig(e) {
-    e.stopPropagation();
-
-    const confirmation = confirm('The spec will be formatted on merge.');
-    if (!confirmation) {
-      return;
-    }
-
-    if (this.props.configEditorString === '{}') {
-      this.props.parseSpec(true);
-      return;
-    }
-
-    try {
-      const spec = JSON.parse(this.props.editorString);
-      const config = JSON.parse(this.props.configEditorString);
-      if (spec.config) {
-        spec.config = mergeDeep(config, spec.config);
-      } else {
-        spec.config = config;
-      }
-      this.props.updateEditorString(stringify(spec));
-
-      this.props.clearConfig();
-    } catch (e) {
-      console.warn(e);
-    }
-
-    this.props.parseSpec(true);
-  }
-
-  public handleExtractConfig() {
-    const confirmation = confirm('The spec and config will be formatted.');
-    if (!confirmation) {
-      return;
-    }
-
-    try {
-      const spec = JSON.parse(this.props.editorString);
-      let config = JSON.parse(this.props.configEditorString);
-      if (spec.config) {
-        config = mergeDeep(config, spec.config);
-        delete spec.config;
-        this.props.updateEditorString(stringify(spec));
-        this.props.setConfigEditorString(stringify(config));
-      }
-    } catch (e) {
-      console.warn(e);
-    }
-    this.props.parseSpec(true);
   }
 
   public editVegaSpec() {
@@ -92,37 +36,6 @@ class CompiledSpecDisplayHeader extends React.PureComponent<Props> {
           <button onClick={this.editVegaSpec} style={{ cursor: 'pointer' }}>
             Edit Vega Spec
           </button>
-
-          {this.props.sidePaneItem === SIDEPANE.Config && (
-            <div className="merge-toolbar">
-              <button>
-                <ArrowUpCircle
-                  data-tip
-                  data-for="mergeConfig"
-                  onClick={e => {
-                    e.stopPropagation();
-                    this.handleMergeConfig(e);
-                  }}
-                />
-              </button>
-              <button>
-                <ArrowDownCircle
-                  data-tip
-                  data-for="extractConfig"
-                  onClick={e => {
-                    e.stopPropagation();
-                    this.handleExtractConfig();
-                  }}
-                />
-              </button>
-              <ReactTooltip id="mergeConfig" effect="solid">
-                <span style={{ textTransform: 'none' }}>Merge config into spec</span>
-              </ReactTooltip>
-              <ReactTooltip id="extractConfig" effect="solid">
-                <span style={{ textTransform: 'none' }}>Extract config from spec</span>
-              </ReactTooltip>
-            </div>
-          )}
 
           <ChevronDown />
         </div>
@@ -146,12 +59,6 @@ class CompiledSpecDisplayHeader extends React.PureComponent<Props> {
 function mapStateToProps(state, ownProps) {
   return {
     compiledVegaSpec: state.compiledVegaSpec,
-    configEditorString: state.configEditorString,
-    editorString: state.editorString,
-    manualParse: state.manualParse,
-    mode: state.mode,
-    sidePaneItem: state.sidePaneItem,
-    themeName: state.themeName,
     value: state.vegaSpec,
   };
 }
@@ -160,13 +67,7 @@ export function mapDispatchToProps(dispatch: Dispatch<EditorActions.Action>) {
   return bindActionCreators(
     {
       clearConfig: EditorActions.clearConfig,
-      parseSpec: EditorActions.parseSpec,
-      setConfig: EditorActions.setConfig,
-      setConfigEditorString: EditorActions.setConfigEditorString,
-      setSidePaneItem: EditorActions.setSidePaneItem,
-      setThemeName: EditorActions.setThemeName,
       toggleCompiledVegaSpec: EditorActions.toggleCompiledVegaSpec,
-      updateEditorString: EditorActions.updateEditorString,
       updateVegaSpec: EditorActions.updateVegaSpec,
     },
     dispatch
