+++ src/components/input-panel/spec-editor-header/index.tsx
@@ -1,12 +1,13 @@
 import stringify from 'json-stringify-pretty-compact';
 import * as React from 'react';
-import { ChevronDown, ChevronUp } from 'react-feather';
+import { ArrowDownCircle, ArrowUpCircle } from 'react-feather';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-
+import ReactTooltip from 'react-tooltip';
 import { bindActionCreators, Dispatch } from 'redux';
+import { mergeDeep } from 'vega-lite/build/src/util';
 import * as EditorActions from '../../../actions/editor';
-import { Mode, SIDEPANE } from '../../../constants';
+import { SIDEPANE } from '../../../constants';
 import ConfigEditorHeader from '../../config-editor/config-editor-header';
 
 const toggleStyle = {
@@ -16,6 +17,58 @@ const toggleStyle = {
 type Props = ReturnType<typeof mapStateToProps> & ReturnType<typeof mapDispatchToProps> & { history: any };
 
 class SpecEditorHeader extends React.PureComponent<Props> {
+  public handleMergeConfig(e) {
+    e.stopPropagation();
+
+    const confirmation = confirm('The spec will be formatted on merge.');
+    if (!confirmation) {
+      return;
+    }
+
+    if (this.props.configEditorString === '{}') {
+      this.props.parseSpec(true);
+      return;
+    }
+
+    try {
+      const spec = JSON.parse(this.props.editorString);
+      const config = JSON.parse(this.props.configEditorString);
+      if (spec.config) {
+        spec.config = mergeDeep(config, spec.config);
+      } else {
+        spec.config = config;
+      }
+      this.props.updateEditorString(stringify(spec));
+
+      this.props.clearConfig();
+    } catch (e) {
+      console.warn(e);
+    }
+
+    this.props.parseSpec(true);
+  }
+
+  public handleExtractConfig() {
+    const confirmation = confirm('The spec and config will be formatted.');
+    if (!confirmation) {
+      return;
+    }
+
+    try {
+      const spec = JSON.parse(this.props.editorString);
+      let config = JSON.parse(this.props.configEditorString);
+      if (spec.config) {
+        config = mergeDeep(config, spec.config);
+        delete spec.config;
+        this.props.updateEditorString(stringify(spec));
+        this.props.setConfigEditorString(stringify(config));
+      }
+    } catch (e) {
+      console.warn(e);
+    }
+    this.props.parseSpec(true);
+  }
+
   public render() {
     const toggleStyleUp = Object.assign({}, toggleStyle, {
       position: 'static',
@@ -51,6 +104,35 @@ class SpecEditorHeader extends React.PureComponent<Props> {
         </ul>
 
         {this.props.sidePaneItem === SIDEPANE.Config && <ConfigEditorHeader />}
+
+        <div className="merge-toolbar">
+          <button>
+            <ArrowUpCircle
+              data-tip
+              data-for="mergeConfig"
+              onClick={e => {
+                e.stopPropagation();
+                this.handleMergeConfig(e);
+              }}
+            />
+          </button>
+          <button>
+            <ArrowDownCircle
+              data-tip
+              data-for="extractConfig"
+              onClick={e => {
+                e.stopPropagation();
+                this.handleExtractConfig();
+              }}
+            />
+          </button>
+          <ReactTooltip id="mergeConfig" effect="solid">
+            <span style={{ textTransform: 'none' }}>Merge config into spec</span>
+          </ReactTooltip>
+          <ReactTooltip id="extractConfig" effect="solid">
+            <span style={{ textTransform: 'none' }}>Extract config from spec</span>
+          </ReactTooltip>
+        </div>
       </div>
     );
   }
@@ -59,8 +141,12 @@ class SpecEditorHeader extends React.PureComponent<Props> {
 function mapStateToProps(state, ownProps) {
   return {
     compiledVegaSpec: state.compiledVegaSpec,
+    configEditorString: state.configEditorString,
+    editorString: state.editorString,
+    manualParse: state.manualParse,
     mode: state.mode,
     sidePaneItem: state.sidePaneItem,
+    themeName: state.themeName,
     value: state.vegaSpec,
   };
 }
@@ -69,8 +155,13 @@ export function mapDispatchToProps(dispatch: Dispatch<EditorActions.Action>) {
   return bindActionCreators(
     {
       clearConfig: EditorActions.clearConfig,
+      parseSpec: EditorActions.parseSpec,
+      setConfig: EditorActions.setConfig,
+      setConfigEditorString: EditorActions.setConfigEditorString,
       setSidePaneItem: EditorActions.setSidePaneItem,
+      setThemeName: EditorActions.setThemeName,
       toggleCompiledVegaSpec: EditorActions.toggleCompiledVegaSpec,
+      updateEditorString: EditorActions.updateEditorString,
       updateVegaSpec: EditorActions.updateVegaSpec,
     },
     dispatch
